* 6.7.8_Restrictions_On_who_Can_Delete_Attachments:
  Implemented Phase 6 Sprint 7 todo item 6.7.8 - Restriction on the deletion of attachments.

Conflicts:
	Configuration/DBScript/Phase6Sprint7.txt

+++ idas.chox.core/src/main/java/idas/chox/core/model/Claim.java
@@ -749,7 +749,7 @@ public class Claim extends Entity implements Serializable {
 
     public Notification GetNotificationById(int id) {
         for (Notification notification : notifications) {
-            if (notification.getId() == id) {
+            if (notification.getId().intValue() == id) {
                 return notification;
             }
         }
